# Implement `slider_iteration()` and give access to it from `slide()`