… NAMESPACE
… man/slider_iteration.Rd