Closes #80 

Implements a context dependent `slider_iteration()` that gives the current iteration while sliding.

- Essentially no overhead when not called.

- Errors when not called from a `slide_*()` function.

- Tries to be as fast as possible by calling up to R but then immediately dropping back to C. I believe this is the approach that data.table takes with GForce.

- The previous point means that users can call it from their own functions without any issues.

It is much faster than the current dev dplyr approach, which currently depends on R level environments. Accessing into those repeatedly is rather slow.

```r
library(slider)

# 30,000 calls to `slider_iteration()`
x <- 1:3e4 + 0L

fn <- function(x) {
  1
}

fn2 <- function(x) {
  slider_iteration()
}

# Master
bench::mark(slide(x, fn), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 slide(x, fn)   8.36ms   10.8ms      93.7    4.01MB     38.1

# This PR (no performance loss when not used!)
bench::mark(slide(x, fn), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 slide(x, fn)   8.31ms     11ms      91.9    4.03MB     36.8

# This PR (~2x slower when used, which isn't that bad, I think)
bench::mark(slide(x, fn2), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 slide(x, fn2)   15.7ms   23.2ms      44.3     341KB     62.7
```

Compare with dev dplyr and `n()`. This is sort of similar, we make 30k groups and call `n()` 30k times.


```r
library(dplyr)

df <- data.frame(x = x)
df <- group_by(df, x)

bench::mark(mutate(df, y = fn()))
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 mutate(df, y = fn())   26.6ms   27.7ms      34.9    2.55MB     14.5

bench::mark(mutate(df, y = n()))
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 mutate(df, y = n())    276ms    288ms      3.47    1.28MB     20.8
```