@DavisVaughan 2020-04-05
@EmilHvitfeldt what do you think? This would help with https://github.com/DavisVaughan/slider/issues/74, you'd access the current index value with `~ date[[slider_iteration()]]`, so essentially it keeps you from needing the `<<-`